hometask48

+++ hometask47-bl-hotel-control/src/main/java/org/spring/hometask47blhotelcontrol/service/HotelService.java
@@ -3,11 +3,13 @@ package org.spring.hometask47blhotelcontrol.service;
 import lombok.RequiredArgsConstructor;
 import org.spring.hometask47blhotelcontrol.dto.HotelDto;
 import org.spring.hometask47blhotelcontrol.entity.HotelEntity;
+import org.spring.hometask47blhotelcontrol.exception.CommonError;
 import org.spring.hometask47blhotelcontrol.mapper.HotelMapper;
 import org.spring.hometask47blhotelcontrol.repository.HotelRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -17,11 +19,19 @@ public class HotelService {
     private final HotelMapper mapper;
 
     public List<HotelDto> getAll(){
-        return mapper.toDtos(repository.findAll());
+        List<HotelEntity> all = repository.findAll();
+        if(all.isEmpty()){
+            throw new CommonError("8081002");
+        }
+        return mapper.toDtos(all);
     }
 
     public HotelDto delete(UUID id){
-        HotelEntity hotelEntity = repository.findById(id).get();
+        Optional<HotelEntity> byId = repository.findById(id);
+        if(byId.isEmpty()){
+            throw new CommonError("808101");
+        }
+        HotelEntity hotelEntity = byId.get();
         repository.delete(hotelEntity);
         return mapper.toDto(hotelEntity);
     }
